Remove hooks, merge ownership and cardstate
